ai-engin/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
python-dotenv
pydantic
openai
pypdf
chromadb
langchain-core